Migrate DefaultLifecycleObserver setup to androidx.startup
